Exception 1: Firebase query returns an iterable of children (DataRecord) objects which match the description of the query. Before this, a map was being used to fetch such results, which was resulting in an exception.
Exception 2:
Due to the DataRecords being fetched inside the main UI thread, the application was crashing as too much work was being done on the main thread and was blocking the app, thereby resulting in an eviction.

Exception 3:
LocationDataRecord did not have a default constructor, thereby making it impossible for firebase to initalize a new record for JSON received.

This change:
1. uses an AsynTask to fetch previous records from Firebase, thereby letting the main UI thread do it's work
2. uses an Iterable to go over the list of records returned by the query
3. adds a default constructor to LocationDataRecord