Change is three-fold.
1.
The use case for NotificationManager expects a Notification to have a:
  a. creationTime (time at which the notification was actually shown to the user, and not when
      the object came into existence)
  b. clickTime (time at which the notification was tapped/clicked) by the user.

This helps the researchers find coorelations between type of notifications, duration, time of day
and user responsiveness.

 2.
 The Notifications use case also expects that if there are "similar" notifications being trigerred, the NotificationManager must not show it to the user. The definition of similar is fuzzy and hence the application developer must have control of what notifications should be categozied as similar.
 This change adds a "Category" field to the NotificaitonEvent. The NotificationManager will ignore any subsequeny notificaitons which share a category with an existing notification in the registry.

3.
As a result of #1 (in the list above), to actually record the clickTime for a notification, the activity that is shown as a result of notification tap must hold the responsibility of letting the notification manager know when a notification was clicked.
This is somewhat roundabout and unintuitive way of doing this, but these are technical limitations we cannot overcome, unless Android API provides a way for developers to add event handlers for notification clicks and dismissal.
This change adds a NotificationClickedEvent trigger to the onCreate method of the BaseActivity, an Activity designed to be extended by any activity alive in a Minuku based app.